Implements MetalWindow::new
