+++ gui/src/graphics/metal/screen.rs
@@ -1,58 +1,33 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 use crate::graphics::Screen;
-use crate::vmm::VmmScreen;
-use metal::{CAMetalLayer, Device, MetalLayer};
-use objc::runtime::{Object, NO, YES};
-use objc::{msg_send, sel, sel_impl};
-use std::error::Error;
-use std::ptr::null_mut;
+use metal::{Device, MetalLayer};
 use thiserror::Error;
 
 /// Implementation of [`Screen`] using Metal.
 ///
 /// Fields in this struct need to be dropped in a correct order.
 pub struct MetalScreen {
-    view: *mut Object,
-    layer: MetalLayer,
     device: Device,
 }
 
 impl MetalScreen {
     pub fn new() -> Result<Self, MetalError> {
-        todo!()
-    }
-
-    pub fn from_screen(screen: &VmmScreen) -> Result<Self, MetalError> {
         // Get Metal device.
         let device = match Device::system_default() {
             Some(v) => v,
             None => return Err(MetalError::GetDeviceFailed),
         };
 
-        // Setup Metal layer.
+        Ok(Self { device })
+    }
+
+    /// # Safety
+    /// The returned [`MetalLayer`] must be dropped before this [`MetalScreen`].
+    pub unsafe fn create_layer(&self) -> MetalLayer {
         let layer = MetalLayer::new();
 
-        layer.set_device(&device);
-
-        // Set view layer.
-        let view = screen.view as *mut Object;
-
-        let _: () = unsafe { msg_send![view, setLayer:layer.as_ref()] };
-        let _: () = unsafe { msg_send![view, setWantsLayer:YES] };
-
-        Ok(Self {
-            view,
-            layer,
-            device,
-        })
-    }
-}
-
-impl Drop for MetalScreen {
-    fn drop(&mut self) {
-        let l: *mut CAMetalLayer = null_mut();
-        let _: () = unsafe { msg_send![self.view, setWantsLayer:NO] };
-        let _: () = unsafe { msg_send![self.view, setLayer:l] };
+        layer.set_device(&self.device);
+        layer
     }
 }
