+++ gui/src/graphics/metal/window.rs
@@ -1,12 +1,56 @@
+use super::MetalScreen;
 use crate::rt::{Hook, RuntimeWindow};
+use metal::{CAMetalLayer, MetalLayer};
+use objc::runtime::{Object, NO, YES};
+use objc::{msg_send, sel, sel_impl};
+use rwh05::{HasRawWindowHandle, RawWindowHandle};
 use std::error::Error;
+use std::ptr::null_mut;
+use std::rc::Rc;
+use std::sync::Arc;
 use winit::dpi::{PhysicalPosition, PhysicalSize};
 use winit::event::{DeviceId, ElementState, InnerSizeWriter, MouseButton, StartCause};
 use winit::event_loop::ControlFlow;
-use winit::window::WindowId;
+use winit::window::{Window, WindowId};
 
 /// Implementation of [`RuntimeWindow`] and [`Hook`] for Metal.
-pub struct MetalWindow {}
+pub struct MetalWindow {
+    view: *mut Object,
+    layer: MetalLayer,
+    window: Window,
+    screen: Arc<MetalScreen>,
+}
+
+impl MetalWindow {
+    pub fn new(
+        screen: &Arc<MetalScreen>,
+        window: Window,
+    ) -> Result<Rc<Self>, Box<dyn Error + Send + Sync>> {
+        let layer = unsafe { screen.create_layer() };
+        let view = match window.raw_window_handle() {
+            RawWindowHandle::AppKit(v) => v.ns_view as *mut Object,
+            _ => unreachable!(),
+        };
+
+        let _: () = unsafe { msg_send![view, setLayer:layer.as_ref()] };
+        let _: () = unsafe { msg_send![view, setWantsLayer:YES] };
+
+        Ok(Rc::new(Self {
+            view,
+            layer,
+            window,
+            screen: screen.clone(),
+        }))
+    }
+}
+
+impl Drop for MetalWindow {
+    fn drop(&mut self) {
+        let l: *mut CAMetalLayer = null_mut();
+        let _: () = unsafe { msg_send![self.view, setWantsLayer:NO] };
+        let _: () = unsafe { msg_send![self.view, setLayer:l] };
+    }
+}
 
 impl RuntimeWindow for MetalWindow {
     fn on_resized(&self, new: PhysicalSize<u32>) -> Result<(), Box<dyn Error + Send + Sync>> {
